move fragment_tree module to its own folder
